Rename typos module
